Updated Value objects
Since the fighter configuration will be treated as a simple value object in the application-specific code, I had to go in and add the value objects back, as well as add duplication methods for each of them. This is only at the fighter level. The hitbox, attack, and window configs still need the same treatment.